requirements.txt
attrs==22.2.0
automium==0.2.6
automium-web==0.1.1
bcrypt==4.0.1
blinker==1.5
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
legacy==0.1.6
MarkupSafe==2.1.2
packaging==23.0
pluggy==1.0.0
pytest==7.2.1
python-decouple==3.7
python-dotenv==0.21.1
SQLAlchemy==2.0.1
typing_extensions==4.4.0
Werkzeug==2.2.2
WTForms==3.0.1
